+++ test/unit/elastic_class/unit_test_run.sh
@@ -23,6 +23,7 @@ echo "Unit test:  elastic_class"
 /usr/bin/python ./test/unit/elastic_class/get_shards.py
 /usr/bin/python ./test/unit/elastic_class/is_active.py
 /usr/bin/python ./test/unit/elastic_class/elasticsearch_connect.py
+/usr/bin/python ./test/unit/elastic_class/elasticsearch_get_dump_list.py
 /usr/bin/python ./test/unit/elastic_class/elasticsearch_is_active.py
 /usr/bin/python ./test/unit/elastic_class/elasticsearch_get_info.py
 /usr/bin/python ./test/unit/elastic_class/elasticsearch_get_master_name.py