+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - elastic_class.ElasticSearch: Added get_nodes method.
 - elastic_class.ElasticSearch: Added get_master_name method.
 - elastic_class.ElasticSearch: Added get_info method.
+- elastic_class.ElasticSearch: Added get_dump_list method.
 
 ### Deprecated
 - elastic_class.is_active function.
@@ -21,6 +22,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - elastic_class.get_nodes function.
 - elastic_class.get_master_name function.
 - elastic_class.get_info function.
+- elastic_class.get_dump_list function.
 - elastic.py: The port and scheme entries in the configuration file.